Separate NFoldManifold and NFoldMetric in nfold_manifold.py, bring ProductRiemannianMetric and ProductManifold together in product_manifold.py
