+++ geomstats/information_geometry/normal.py
@@ -17,7 +17,7 @@ from geomstats.geometry.poincare_half_space import (
     PoincareHalfSpaceMetric,
 )
 from geomstats.geometry.product_manifold import ProductManifold
-from geomstats.geometry.product_riemannian_metric import NFoldMetric
+from geomstats.geometry.nfold_manifold import NFoldMetric
 from geomstats.geometry.pullback_metric import PullbackDiffeoMetric
 from geomstats.geometry.riemannian_metric import RiemannianMetric
 from geomstats.geometry.spd_matrices import SPDAffineMetric, SPDMatrices
